Moved dump-related stuff into a separate management script.
