+++ acousticbrainz/utils.py
@@ -1,3 +1,4 @@
+import errno
 import json
 import os
 
@@ -100,3 +101,13 @@ def interpret_high_level(hl):
     other.append(_interpret("ISMIR04 Rhythm", hl['highlevel']['ismir04_rhythm'], .6))
 
     return genres, moods, other
+
+
+def create_path(path):
+    """Creates a directory structure if it doesn't exist yet."""
+    try:
+        os.makedirs(path)
+    except OSError as exception:
+        if exception.errno != errno.EEXIST:
+            raise Exception("Failed to create directory structure %s. Error: %s" %
+                            (path, exception))
